requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
dj-database-url==2.1.0
pytz==2023.3.post1
sqlparse==0.4.4
virtualenv==20.24.5
psycopg==3.1.12
psycopg2==2.9.9
gunicorn==21.2.0
Pillow==10.0.1
psycopg2-binary==2.9.9
pytz==2023.3.post1
whitenoise==6.5.0
python-dotenv==1.0.0
django-cors-headers==4.2.0
